+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Nodes can be found in direct children
+
 ## [1.0.8] - 2024-05-10
 
 ### Added
@@ -74,6 +80,7 @@
 - 'Github Actions Pipeline'
 - 'Github Actions Pipeline: Add SDK file containing flutter into .github/workflows to make github installing flutter and not dart SDK'
 
+[Unreleased]: https://github.com/inlavigo/supply_chain/compare/1.0.8...HEAD
 [1.0.8]: https://github.com/inlavigo/supply_chain/compare/1.0.7...1.0.8
 [1.0.7]: https://github.com/inlavigo/supply_chain/compare/1.0.6...1.0.7
 [1.0.6]: https://github.com/inlavigo/supply_chain/compare/1.0.5...1.0.6